+++ docs/changes.md
@@ -7,8 +7,9 @@
 
 - Fix meshlab interface thanks to @JeffreyWardman in #924
 - Update `Slicer3DPlotter` thanks to @daniel-a-diaz in #925
-- Improvements on `applications.Browser`
 - Improvemnets on `applications.Slicer3DPlotter`
+- Improvements on `applications.Browser`
+- add background radial gradients
 
 
 ### Breaking changes
@@ -21,6 +22,7 @@
 ```
 examples/advanced/timer_callback1.py
 examples/advanced/timer_callback2.py
+examples/advanced/interpolate_scalar5.py
 examples/basic/buttons.py
 examples/basic/input_box.py
 examples/basic/sliders2.py